Changing this for demonstration
Created a basic template we can build from, feel free to add or delete as you see necessary, just whipped it up quick, we may want to change some things such as how I currently have owner and user as two different arrays of objects, we may want to change this so that it is simply users, and they have an option to post listings, such as something like Kijiji? Or even have an additional field in the object to state whether they made a user account or owner account? Just things to think about.