Correction of errors in HMTL and JS.  Reconstruction of the Owners property
Colin's code was pushed before my code, though in the meeting we talked I needed to upload my code first.   This implied my code was taken over the existing one on the main branch (outdated).

The HTML code is a compilation of exercises from class. My code was not uploaded due to the conflicts my corrections were presenting against the incorrect code.  (JS was pointing to a class exercise, dropdown list code is using non related names, the identification of the object is pointing to non existing objects and is not considering other elements on the document, the CS is not allowing the SignUp page to display since its from another exercise).

I need to include the final code for the pushing to object, I had to create all the code again.